requirements.txt
# requirements

discord
python-dotenv
json
requests